+++ diffsky/experimental/kernels/tests/test_dbk_specphot_kernels.py
@@ -1,0 +1,15 @@
+""""""
+
+from .. import dbk_specphot_kernels as dbkspk
+
+
+def test_mc_dbk_phot_kern():
+    pass
+
+
+def test_mc_dbk_specphot_kern():
+    pass
+
+
+def test_mc_lc_dbk_sed_kern():
+    pass
